+++ src/App/VRChatContentPublisher.App/ViewModels/Settings/DebugSettingsViewModel.cs
@@ -7,24 +7,10 @@ using VRChatContentPublisher.Core.Settings.Models;
 namespace VRChatContentPublisher.App.ViewModels.Settings;
 
 public sealed partial class DebugSettingsViewModel(
-    NetworkDiagnostic.NetworkDiagnosticWindowViewModel networkDiagnosticWindowViewModel,
-    IWritableOptions<AppSettings> appSettings
+    NetworkDiagnostic.NetworkDiagnosticWindowViewModel networkDiagnosticWindowViewModel
 ) : ViewModelBase
 {
     public string LogsPath => AppStorageService.GetLogsPath();
-
-    public bool UseRgbCyclingBackgroundMenu
-    {
-        get;
-        set
-        {
-            if (field == value)
-                return;
-
-            OnPropertyChanging();
-            appSettings.UpdateAsync(settings => { settings.UseRgbCyclingBackgroundMenu = value; });
-        }
-    } = appSettings.Value.UseRgbCyclingBackgroundMenu;
 
     [RelayCommand]
     private void OpenNetworkDiagnosticWindow()